+++ lib/amazing_print/inspector.rb
@@ -159,13 +159,23 @@ module AmazingPrint
       @options.merge!(options)
     end
 
+    def find_dotfile
+      xdg_config_home = File.expand_path(ENV.fetch('XDG_CONFIG_HOME', '~/.config'))
+      xdg_config_path = File.join(xdg_config_home, 'aprc')  # ${XDG_CONFIG_HOME}/aprc
+      
+      return xdg_config_path if File.exist?(xdg_config_path)
+      
+      # default to ~/.aprc
+      File.join(ENV['HOME'], '.aprc')
+    end
+
     # This method needs to be mocked during testing so that it always loads
     # predictable values
     #---------------------------------------------------------------------------
     def load_dotfile
       return if @@dotfile # Load the dotfile only once.
 
-      dotfile = File.join(ENV['HOME'], '.aprc')
+      dotfile = find_dotfile
       load dotfile if dotfile_readable?(dotfile)
     end